Bug 10078 New behaviour giving credit for parts that were right on earlier tries.
This is a (non-archetypal) variant of the interactive behaviour that
a question type can create in the make_behaviour method in place of
the interactive behaviour.

An easy way to do this is to extend question_graded_automatically_with_countback
If you can't do that for some reason, just implement the
question_automatically_gradable_with_countback interface and use
your own make_behaviour method.

That interface requries a new method
compute_final_grade($responses, $totaltries) that is passed the
complete response history, and has to compute the final grade.
This is called when the question attempt is over.

The new scoring has been implemented for the match, ddwtos and
oumultiresponse question types.