Bug 10468 and 10467 more consistent display of colour for feedback on the student's response.
The colour is now linked to the correctness (Whether correct) display option.

It is now only shown to indicate whether (parts of) the student's response was correct, not to also indicate what the right answer is.

The coloured background, and the tick/cross icon, are now either shown together, or not at all.