Bug 11098 only upgrade question attempts for summative quizzes or open courses. This change also block access to parts of quizzes that depend on attempt data, if not all attempts have been converted.
A summative quiz is defined as one with an idnumber.
An open course is defined as one where either there is no row in the vl_v_crs_version_pres, or there is a row there with a vle_student_close_date that is blank or after 2010-12-01
If the vl_v_crs_version_pres view does not exist, all quizzes are upgraded.
Also, skip the single big course MU123-10B. We will convert that manually shortly after the main upgrade.